requirements.txt
accelerate==1.6.0
arch==7.2.0
datasets==3.5.0
einops==0.8.1
huggingface_hub==0.30.2
mamba_ssm==2.2.4
matplotlib==3.10.3
nltk==3.9.1
numpy==2.2.6
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
reformer_pytorch==1.4.4
scikit_learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
statsmodels==0.14.4
sympy==1.13.3
torch==2.8.0.dev20250422+cu128
torchinfo==1.8.0
tqdm==4.67.1
